404 Error has been fixed
I found the problem. There are 2 index.htmls to refer to, and GitHub pages is using the wrong one (the one in project instead of the one in dist).

This commit adds a command to the deploy.yml, which instructs GitHub Pages to not use Jekyll (which can sometimes interfere with how files are processed).